AiDecompiler/requirements.txt
diskcache==5.6.3
Jinja2==3.1.5
llama_cpp_python==0.3.7
MarkupSafe==3.0.2
numpy==2.2.2
typing_extensions==4.12.2
